requirements.txt
# Core bot framework
python-telegram-bot==22.3
python-telegram-bot[job-queue]==22.3

# HTTP requests
requests>=2.31.0

# Data processing and analysis
numpy>=1.24.0
pandas>=2.0.0

# Timezone handling
pytz>=2023.3

# Caching
cachetools>=5.3.0

# Plotting and visualization
matplotlib>=3.7.0
seaborn>=0.12.0